Splitting common code shared with the bootloader
After upgrading the version of the arduino-pico library
the bootloader in debug mode was about 1.5k too large.

This method is a bit extreme as it splits the code into what is common
between the bootloader and the main RP2040 code out into it's own
library. This is to decrease what gets linked when the bootloader is
compiled.

There is probably a way to do this without splitting to code
as the main difference between the this and the
commit before it is upgrading to GCC 14, the pico-sdk and
the arduino-pico library now have support for the RP235XX line.
