+++ boards/zuluide_rp2040.json
@@ -9,7 +9,7 @@
     },
     "core": "earlephilhower",
     "cpu": "cortex-m0plus",
-    "extra_flags": "-D ARDUINO_RHC_ZULUIDE_RP2040 -DARDUINO_ARCH_RP2040 -DUSBD_MAX_POWER_MA=500",
+    "extra_flags": "-DARDUINO_GENERIC_RP2040 -DARDUINO_ARCH_RP2040 -DUSBD_MAX_POWER_MA=500 ",
     "f_cpu": "125000000L",
     "hwids": [
       [
@@ -22,7 +22,7 @@
       ]
     ],
     "mcu": "rp2040",
-    "variant": "zuluide_rp2040"
+    "variant": "generic"
   },
   "debug": {
     "jlink_device": "RP2040_M0_0",
@@ -52,5 +52,5 @@
     ]
   },
   "url": "https://www.zuluide.com",
-  "vendor": "Rabbit Hole Computing"
+  "vendor": "Rabbit Hole Computing LLC"
 }
