+++ lib/RP2040_common/rp2040_sdio.cpp
@@ -35,8 +35,9 @@
 #include <hardware/pio.h>
 #include <hardware/dma.h>
 #include <hardware/gpio.h>
-#include <ZuluIDE_platform.h>
 #include <ZuluIDE_log.h>
+#include <ZuluIDE_platform_gpio.h>
+#include <string.h>
 
 #define SDIO_PIO pio1
 #define SDIO_CMD_SM 0
@@ -708,7 +709,8 @@ static void rp2040_sdio_tx_irq()
 // Check if transmission is complete
 sdio_status_t rp2040_sdio_tx_poll(uint32_t *bytes_complete)
 {
-    if (SCB->ICSR & SCB_ICSR_VECTACTIVE_Msk)
+    // SCB_ICSR_VECTACTIVE_Msk (0x1FFUL)
+    if (scb_hw->icsr & (0x1FFUL))
     {
         // Verify that IRQ handler gets called even if we are in hardfault handler
         rp2040_sdio_tx_irq();
@@ -784,6 +786,8 @@ void rp2040_sdio_init(int clock_divider)
     static bool resources_claimed = false;
     if (!resources_claimed)
     {
+        pio_sm_unclaim(SDIO_PIO, SDIO_CMD_SM);
+        pio_sm_unclaim(SDIO_PIO, SDIO_DATA_SM); 
         pio_sm_claim(SDIO_PIO, SDIO_CMD_SM);
         pio_sm_claim(SDIO_PIO, SDIO_DATA_SM);
         dma_channel_claim(SDIO_DMA_CH);
