+++ lib/ZuluControl/include/zuluide/images/image.h
@@ -22,6 +22,7 @@
 #pragma once
 
 #include <string>
+#include <stdint.h>
 
 namespace zuluide::images {
