+++ platformio.ini
@@ -2,11 +2,12 @@
 default_envs = ZuluIDE_RP2040
 
 [env:ZuluIDE_RP2040]
-platform = https://github.com/maxgerhardt/platform-raspberrypi.git#5e87ae34ca025274df25b3303e9e9cb6c120123c
+platform = https://github.com/maxgerhardt/platform-raspberrypi.git#19e30129fb1428b823be585c787dcb4ac0d9014c
 board_build.core = earlephilhower
 platform_packages =
-    framework-arduinopico@https://github.com/rabbitholecomputing/arduino-pico.git#v3.7.2-ZuluIDE
+    framework-arduinopico@https://github.com/rabbitholecomputing/arduino-pico.git#v4.3.0-DaynaPORT
 framework = arduino
+lib_ldf_mode = chain+
 board = zuluide_rp2040
 extra_scripts = src/build_bootloader.py
 board_build.ldscript = lib/ZuluIDE_platform_RP2040/rp2040.ld
@@ -16,11 +17,13 @@ lib_deps =
     minIni
     ZuluControl
     ZuluI2S
+    RP2040_common
+    ZuluIDE_platform_RP2040
     CUEParser=https://github.com/rabbitholecomputing/CUEParser
     adafruit/Adafruit GFX Library
     adafruit/Adafruit BusIO
-    ZuluIDE_platform_RP2040
     adafruit/Adafruit SSD1306
+    Wire
 debug_tool = cmsis-dap
 debug_build_flags =
     -O2 -ggdb -g3
@@ -36,3 +39,4 @@ build_flags =
     -DPLATFORM_MASS_STORAGE
     -DSSD1306_NO_SPLASH
     -DENABLE_AUDIO_OUTPUT
+
